+++ src/reuse/__init__.py
@@ -22,12 +22,16 @@ import re
 from dataclasses import dataclass, field
 from typing import NamedTuple, Set
 
+try:
+    from importlib.metadata import PackageNotFoundError, version
+except ImportError:
+    from importlib_metadata import PackageNotFoundError, version
+
 from boolean.boolean import Expression
-from pkg_resources import DistributionNotFound, get_distribution
 
 try:
-    __version__ = get_distribution(__name__).version
-except DistributionNotFound:
+    __version__ = version("reuse")
+except PackageNotFoundError:
     # package is not installed
     __version__ = "1.1.2"